backend/requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
psycopg2-binary
boto3
django-storages
Pillow